Extract "project review" buttons
