+++ front/app/containers/Admin/projects/project/projectHeader/PublicationButtons/ReviewFlow/ReviewRequestButton/index.tsx
@@ -5,8 +5,8 @@ import { Button, Tooltip } from '@citizenlab/cl2-component-library';
 import { trackEventByName } from 'utils/analytics';
 import { useIntl } from 'utils/cl-intl';
 
-import messages from '../messages';
-import tracks from '../tracks';
+import messages from '../../messages';
+import tracks from '../../tracks';
 
 import ReviewRequestDropdown from './ReviewRequestDropdown';
