+++ front/app/containers/Admin/projects/project/projectHeader/PublicationButtons/index.tsx
@@ -2,7 +2,6 @@ import React from 'react';
 
 import { Button, Tooltip, Box } from '@citizenlab/cl2-component-library';
 
-import useApproveProjectReview from 'api/project_reviews/useApproveProject';
 import useProjectReview from 'api/project_reviews/useProjectReview';
 import { IProjectData } from 'api/projects/types';
 import useUpdateProject from 'api/projects/useUpdateProject';
@@ -13,33 +12,22 @@ import { useIntl } from 'utils/cl-intl';
 import { usePermission } from 'utils/permissions';
 
 import messages from './messages';
-import ReviewRequestButton from './ReviewRequestButton';
+import ReviewFlow from './ReviewFlow';
 
 const PublicationButtons = ({ project }: { project: IProjectData }) => {
   const isProjectReviewEnabled = useFeatureFlag({ name: 'project_review' });
 
   const { formatMessage } = useIntl();
 
-  const { data: projectReview, isLoading: isProjectReviewLoading } =
-    useProjectReview(project.id);
+  const { data: projectReview } = useProjectReview(project.id);
 
   const { mutate: updateProject, isLoading: isUpdatingProjectLoading } =
     useUpdateProject();
-
-  const {
-    mutate: approveProjectReview,
-    isLoading: isApprovingProjectReviewLoading,
-  } = useApproveProjectReview();
 
   const canPublish = usePermission({
     item: project,
     action: 'publish',
     context: projectReview?.data.attributes.state === 'approved',
-  });
-
-  const canReview = usePermission({
-    item: project,
-    action: 'review',
   });
 
   const publishProject = () => {
@@ -56,13 +44,6 @@ const PublicationButtons = ({ project }: { project: IProjectData }) => {
     return null;
   }
 
-  const approvalPending = projectReview?.data.attributes.state === 'pending';
-  const approvalGranted = projectReview?.data.attributes.state === 'approved';
-
-  const showProjectApprovalButton =
-    isProjectReviewEnabled && approvalPending && canReview;
-  const showReviewRequestButton =
-    isProjectReviewEnabled && !canReview && !approvalGranted;
   const showPublishButton = !isProjectReviewEnabled || canPublish;
 
   return (
@@ -92,36 +73,7 @@ const PublicationButtons = ({ project }: { project: IProjectData }) => {
         </Tooltip>
       )}
 
-      {showReviewRequestButton && (
-        <Box position="relative">
-          <ReviewRequestButton
-            projectId={project.id}
-            inFolder={!!project.attributes.folder_id}
-            approvalPending={approvalPending}
-            processing={isProjectReviewLoading}
-          />
-        </Box>
-      )}
-
-      {showProjectApprovalButton && (
-        <Tooltip
-          content={formatMessage(messages.approveTooltip)}
-          placement="bottom"
-        >
-          <Button
-            buttonStyle="admin-dark"
-            icon="unlock"
-            onClick={() => approveProjectReview(project.id)}
-            processing={isApprovingProjectReviewLoading}
-            size="s"
-            padding="4px 8px"
-            iconSize="20px"
-            id="e2e-approve-project"
-          >
-            {formatMessage(messages.approve)}
-          </Button>
-        </Tooltip>
-      )}
+      {isProjectReviewEnabled && <ReviewFlow project={project} />}
     </Box>
   );
 };
